+++ packages/console/View.tsx
@@ -4,6 +4,11 @@ import { JsLink } from "^jab-react";
 
 import { ConsoleStateProv, ViewEntry, ViewEntryProps } from "./internal";
 
+/**
+ * tobe-deprecated Use ViewProps instead.
+ */
+export type Props = ViewProps;
+
 export type ViewProps = { showClearLink?: boolean } & Pick<
   ConsoleStateProv,
   "logs" | "clearAllLogs"